add convenience API for exposed packages declared by annotation
This makes the use case to expose custom packages for each module via the module descriptor annotation very simple. I.e. users can simply declare a property name of a `String[]` property on the descriptor annotation, and we will treat all the values in this array as package identifiers and only allow access to those classes within the module that match one of the package identifiers.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>